Moved all UI admin features into folder "admin". Created component AdminComponent, which represents all UI for admin.
