+++ MedicalDocManagment.WebUI/app/admin/admin.component.html
@@ -1,0 +1,5 @@
+﻿<usersList>Завантажується список користувачів...</usersList>
+<hr>
+<app-user-add></app-user-add>
+<hr />
+<users-search-form></users-search-form>
