+++ MedicalDocManagment.WebUI/app/admin/admin.component.ts
@@ -1,0 +1,9 @@
+﻿import { Component } from '@angular/core';
+
+@Component({
+    moduleId: module.id,
+    selector: 'admin-component',
+    templateUrl: 'admin.component.html'
+})
+export class AdminComponent {
+}
