+++ MedicalDocManagment.WebUI/app/admin/users-list/page/item-actionlist-delete/confirmation/confirmation.modal.html
