+++ MedicalDocManagment.WebUI/app/admin/users-list/page/item-actionlist-delete/delete.component.html
