+++ MedicalDocManagment.WebUI/app/admin/users-list/page/item-actionlist-edit/modal/confirmation/confirmation.modal.ts
