+++ MedicalDocManagment.WebUI/app/admin/users-list/page/item-actionlist-edit/modal/modal.component.html
