+++ MedicalDocManagment.WebUI/app/admin/users-list/page/item-actionlist.component.ts
@@ -1,7 +1,7 @@
 ﻿import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { Response } from '@angular/http';
 
-import UserModel from '../../models/usermodel'
+import UserModel from '../../../models/usermodel'
 
 @Component({
     moduleId: module.id,
