+++ MedicalDocManagment.WebUI/app/admin/users-list/page/views/item-actionlist.component.html
