+++ MedicalDocManagment.WebUI/app/admin/users-list/page/views/item.component.css
