+++ MedicalDocManagment.WebUI/app/admin/users-list/page/views/page-without-pagination.component.html
