+++ MedicalDocManagment.WebUI/app/app.module.ts
@@ -4,29 +4,31 @@ import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 
 import { AppComponent } from './app.component';
-import { UserAddComponent } from './user-add/user-add.component';
-import { UsersListModule } from './users-list/users-list.module';
-import { UsersListPaginateComponent } from './users-list-paginate/users-list-paginate.component';
-import { UsersSearchFormComponent } from './users-search/users-search-form.component';
+import { AdminComponent } from './admin/admin.component';
+import { UserAddComponent } from './admin/user-add/user-add.component';
+import { UsersListModule } from './admin/users-list/users-list.module';
+import { UsersListPaginateComponent } from './admin/users-list-paginate/users-list-paginate.component';
+import { UsersSearchFormComponent } from './admin/users-search/users-search-form.component';
 import { Ng2PaginationModule } from 'ng2-pagination';
 import { SimpleNotificationsModule, PushNotificationsModule } from 'angular2-notifications';
 
 @NgModule({
-  imports: [
-    BrowserModule,
-    FormsModule,
-    HttpModule,
-    UsersListModule,
-    Ng2PaginationModule,
-    SimpleNotificationsModule,
-    PushNotificationsModule
-  ],
-  declarations: [
-    AppComponent,
-    UserAddComponent,
-    UsersListPaginateComponent,
-    UsersSearchFormComponent
-  ],
-  bootstrap: [AppComponent]
+    imports: [
+        BrowserModule,
+        FormsModule,
+        HttpModule,
+        UsersListModule,
+        Ng2PaginationModule,
+        SimpleNotificationsModule,
+        PushNotificationsModule
+    ],
+    declarations: [
+        AppComponent,
+        AdminComponent,
+        UserAddComponent,
+        UsersListPaginateComponent,
+        UsersSearchFormComponent
+    ],
+    bootstrap: [AppComponent]
 })
 export class AppModule { }
