+++ MedicalDocManagment.WebUI/app/views/app.component.html
@@ -1,6 +1,1 @@
-﻿<usersList>Завантажується список користувачів...</usersList>
-<hr>
-<app-user-add></app-user-add>
-<hr />
-<users-search-form></users-search-form>
-
+﻿<admin-component></admin-component>
